Create README.md
Bank customer churn is a critical issue impacting the financial industry, and accurately identifying potential churners is essential for proactive retention strategies. This project focuses on the application of diverse preprocessing techniques and ensemble learning models to enhance the accuracy of bank churn prediction.
The dataset, consisting of comprehensive customer information, undergoes meticulous preprocessing, including feature scaling, and encoding. To capitalize on the strengths of various algorithms, AdaBoost and XGBoost models are employed. Furthermore, hyperparameter tuning is exclusively applied to the AdaBoost model to maximize its predictive capabilities.
Innovatively, a Voting Classifier is introduced to combine the strengths of individual models, resulting in a robust ensemble model. The culmination of these efforts yields compelling results, with a ROC-AUC score of 87% and an accuracy of 85%. Notably, the AdaBoost model, with tuned hyperparameters, significantly contributes to the ensemble's predictive prowess.